requirements.txt
fastapi==0.110.3
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
